requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
google-generativeai
python-dotenv
cachetools
requests
